Migrate to next.js
