+++ package.json
@@ -17,26 +17,25 @@
     "emoji-picker-react": "^3.5.1",
     "firebase": "^9.6.9",
     "framer-motion": "^6.3.0",
+    "next": "^12.1.6",
     "react": "^17.0.2",
     "react-datepicker": "^4.7.0",
     "react-dom": "^17.0.2",
     "react-firebase-hooks": "^5.0.3",
     "react-hot-toast": "^2.2.0",
     "react-json-to-csv": "^1.0.4",
-    "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
     "recharts": "^2.1.9",
     "sass": "^1.49.9",
     "typescript": "^4.6.2",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --watchAll=false",
-    "test-ci": "cross-env CI=1 react-scripts test --watchAll=false",
-    "test-watch": "react-scripts test",
-    "eject": "react-scripts eject",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "test": "jest --watchAll=false",
+    "test-ci": "cross-env CI=1 jest test --watchAll=false",
+    "test-watch": "jest test --watch",
     "lint": "eslint --ext .js,.ts,.tsx src/",
     "lint-fix": "eslint --ext .js,.ts,.tsx src/ --fix"
   },
