+++ src/Components/LoginButton/LoginButton.tsx
@@ -1,3 +1,5 @@
+import Image from 'next/image';
+
 interface LoginButtonProps {
   image: string;
   onClick?: () => void;
@@ -18,7 +20,7 @@ function LoginButton({
       className="flex justify-center py-2 pr-4 pl-2 font-semibold text-zinc-900 bg-white rounded-lg shadow border items-center hover:bg-zinc-50 active:bg-zinc-50 w-full sm:w-auto"
       {...props}
     >
-      <img className="mr-4 ml-2" width="24px" height="24px" src={image} />
+      <Image className="mr-4 ml-2" width="24px" height="24px" src={image} />
       {children}
     </button>
   );
