+++ src/Components/Logo/Logo.tsx
@@ -1,5 +1,5 @@
 import clsx from 'clsx';
-import { Link } from 'react-router-dom';
+import Link from 'next/link';
 
 interface LogoProps {
   classNames?: string;
@@ -7,14 +7,12 @@ interface LogoProps {
 
 function Logo({ classNames = '' }: LogoProps) {
   return (
-    <Link
-      to={'/'}
-      className={clsx(
-        'flex items-center font-serif text-2xl text-zinc-600',
-        classNames
-      )}
-    >
-      Employee Pulse
+    <Link href={'/'} className={'flex items-center'} passHref>
+      <a
+        className={clsx('block font-serif text-2xl text-zinc-600', classNames)}
+      >
+        Employee Pulse
+      </a>
     </Link>
   );
 }
