--- src/Components/ProtectedRoute/index.ts
@@ -1,2 +1,0 @@
-export { default } from './ProtectedRoute';
-export * from './ProtectedRoute';
