--- src/Pages/HomePage/index.ts
@@ -1,2 +1,0 @@
-export { default } from './HomePage';
-export * from './HomePage';
