--- src/Pages/LoginPage/index.ts
@@ -1,2 +1,0 @@
-export { default } from './LoginPage';
-export * from './LoginPage';
