--- src/Pages/SurveyAnswerPage/index.ts
@@ -1,2 +1,0 @@
-export { default } from './SurveyAnswerPage';
-export * from './SurveyAnswerPage';
