--- src/Pages/SurveyCreatePage/index.ts
@@ -1,2 +1,0 @@
-export { default } from './SurveyCreatePage';
-export * from './SurveyCreatePage';
