--- src/Pages/SurveyListPage/index.ts
@@ -1,2 +1,0 @@
-export { default } from './SurveyListPage';
-export * from './SurveyListPage';
