--- src/Pages/SurveyPage/index.ts
@@ -1,2 +1,0 @@
-export { default } from './SurveyPage';
-export * from './SurveyPage';
